Remove data model and move its content to view model.
